+++ GoalMogul-iOS/src/Router.js
@@ -69,7 +69,7 @@ class RouterComponent extends Component {
           <Lightbox>
             <Scene key="root" hideNavBar>
               <Scene key="auth" initial hideNavBar>
-                <Scene key="splash" component={Profile} initial />
+                <Scene key="splash" component={SplashScreen} initial />
                 <Scene key="login" component={LoginPage} />
               </Scene>
 
@@ -119,8 +119,8 @@ class RouterComponent extends Component {
                     hideNavBar
                   >
                     <Scene key="home" component={Home} initial />
-                    <Scene key="profile" component={Profile} panHandlers={null} />
-                    <Scene key="profileDetail" component={ProfileDetail} panHandlers={null} />
+                    <Scene key="profile" component={Profile} />
+                    <Scene key="profileDetail" component={ProfileDetail} />
                     <Scene key="setting" component={Setting} />
                     <Scene key="email" component={Email} />
                     <Scene key="editEmailForm" component={EditEmailForm} />