+++ GoalMogul-iOS/src/reducers/Profile.js
@@ -8,7 +8,9 @@ import {
   PROFILE_UPDATE_SUCCESS,
   PROFILE_UPDATE_FAIL,
   SETTING_EMAIL_UPDATE_SUCCESS,
-  PROFILE_SWITCH_TAB
+  PROFILE_SWITCH_TAB,
+  MEET_UPDATE_FRIENDSHIP_DONE,
+  PROFILE_FETCHING_FAIL
 } from '../actions/types';
 
 export const PROFILE_FETCH_MUTUAL_FRIEND_DONE = 'profile_fetch_mutual_friend_done';
@@ -45,6 +47,7 @@ const INITIAL_STATE = {
   * Ignore if it's self
   */
   friendship: {
+    _id: undefined,
     status: undefined // one of [undefined, 'Invited', 'Accepted']
   },
 
@@ -106,6 +109,9 @@ export default (state = INITIAL_STATE, action) => {
     case PROFILE_OPEN_PROFILE:
       return { ...state, userId: action.payload, loading: true };
 
+    case PROFILE_FETCHING_FAIL:
+      return { ...state, loading: false };
+
     case PROFILE_FETCHING_SUCCESS:
       return { ...state, user: action.payload, loading: false };
 
@@ -149,7 +155,42 @@ export default (state = INITIAL_STATE, action) => {
 
     // profile fetch friendship request done
     case PROFILE_FETCH_FRIENDSHIP_DONE: {
-      return { ...state };
+      let newFriendship = _.cloneDeep(state.friendship);
+      if (action.payload !== undefined && action.payload !== null) {
+        newFriendship = action.payload;
+      }
+      return { ...state, friendship: newFriendship };
+    }
+
+    /**
+    payload: {
+      type: ['acceptFriend', 'deleteFriend', 'requestFriend']
+      tab: ['requests.outgoing', 'requests.incoming', 'friends', 'suggested']
+      data: if 'deleteFriend' or 'acceptFriend', then friendshipId. Otherwise, userId
+    }
+    */
+    case MEET_UPDATE_FRIENDSHIP_DONE: {
+      // If updating current profile's friendship, then update the status
+      const { type, data, message } = action.payload;
+      const { id } = data;
+      const resData = data.data;
+      let newFriendship = _.cloneDeep(state.friendship);
+      if (!message) {
+        // If no message, upate succeed
+        if (type === 'requestFriend' && id === state.userId) {
+          if (resData) {
+            newFriendship = _.cloneDeep(resData);
+          } else {
+            newFriendship.status = 'Invited';
+          }
+        } else if (type === 'deleteFriend' && id === state.friendship._id) {
+          newFriendship.status = undefined;
+        } else if (type === 'acceptFriend' && id === state.friendship._id) {
+          newFriendship.status = 'Accepted';
+        }
+      }
+
+      return { ...state, friendship: newFriendship };
     }
 
     default: